Cherry-pick 3 commits from release/2.6 to dygraph (#8821)
* Fix typo and grammatical error (#8785)

* Fix broken link to install paddlepaddle (#8729)

The link provided for installing paddlepaddle doesn't work. Hence, this change updates the broken link to install paddlepaddle for CPU and GPU.

* Fix: broken link for whl package documentation. (#8719)

This proposed change fixes the broken link under the section `2.4 Parameter Description` in the last line that says:
Most of the parameters are consistent with the PaddleOCR whl package, see `whl package documentation`.